requirements.txt
Pillow>=10
python-dotenv>=1.0.1
spotipy>=2.21.1
yandex-music>=2.0.0
